Added gage data sanity checks, did minor bug fixes
Gage data sanity checks for either completely or partially missing data.  Email to Jason with details:
I've finished my edits, and it appears to be working.  I also made a number of smaller corrections in places where I saw bugs in the code.  Here's a laundry list in case I screwed something up due to misinterpretation:
Near start, I moved the pathSep declaration so it didn't get cleared by the clearvars.
Within the "if useDefaultStationList==false" statement, I've added some comments about changes made there to fix some variables, but it's under a condition I'm not able to test, and so cannot verify functionality.
Near line 212, I changed "if exist('approxres','var')==0" to "==1"
Near line 558, within "if writeoutfil==1", I added the "nondefaultvals=..." line since a definition was missing for that variable.
Some other small changes that shouldn't affect anything.
The robustness modifications are in line with what we'd talked about.  There's a check for gages with no data and a check for gages that're missing some fraction of the data requested.  I didn't do a "big number" check or anything of the sort since, as we'd discussed, this sort of thing is quite rare in my experience with NOAA gages.  I added a new parameter, misspropcutoff, which is the fraction of data that can be missing before a gage is thrown out.  I, without much thought, have set it to 0.01 by default.  I think it's setup so you can do it as a command-line option.

That said, I would encourage a worst-case scenario check be implemented before this system be used in a completely automated fashion.  This would look something like this:
if any(abs(diff(sort(mean([dat(:).wl]))))>diffcutoff)|any(abs(mean([dat(:).wl]))>meancutoff)
    halt offset surface work and produce an offset surface of zeros
end
Where diffcutoff and meancutoff would be user-chosen cutoffs.