Made emailattach.py aware of asgs-global.conf
Issue 95: If $HOME/asgs-global.conf exists, then the script will attempt
to read in the credentials from it. The 'server' and 'fro' (from) params
are also ignored in favor of what's in the configuration file. This change
is minimal and is intended to provide backwards compatible support for
existing installations that utilize 'localhost' for sending emails.