Default base directory of PREPPEDARCHIVE and HINDCASTARCHIVE to SCRATCH
Issue 101: Added a new variable option for the config file to set the base
directory for PREPPEDARCHIVE and HINDCASTARCHIVE. If RUNARCHIVEBASE is not
otherwise set in the configuration file, SCRATCHDIR is used. SCRATCHDIR also
now defaults to $SCRATCH (common envar on HPC systems) unless SCRATCHDIR is
explicitly defined in the asgs_main.sh configuration file (typically defined).