Created a Perl script to manage building all Fortran utilities.
Issue 107: Built an extensible Perl script to manage the building
of an entire ASGS environment locally. Currently the steps in the
script focus on the building of NetCDF, HDF5, and any Fortran
utilities in the ASGS repository that have a makefile associated
with them. Extending this script to handle most, if not all of the
tasks associated with building a working ASGS environment from
scratch on any platform should be broken out into future Issues.
The groundwork has been laid with this script. A basic unit test
was created, but it's not sufficient to exercise the script. Some
changes were made to existing makefiles and scripts that are used
to do the tedious work associate with building and compiling.