Moved where openmpi for fortran was built to avoid conflict. (#124)
Moved where openmpi for fortran was built to avoid conflict.

Issue 123: If openmpi (for gfortan) was built prior to a call using the
"intel" compiler, then the environmental variables were used to build adcirc/
padcirc, the PATH would be set up incorrectly - finding the gfortran
mpi compiler wrappers before the proper intel wrappers that are provided
for on the platform. Also cleaned up some of the program flow for
when --dump-env was used. Added "LIBRARY_PATH" to the netcdf/hdf5 step,
which is just a duplicate of LD_LIBRARY_PATH but used by the intel compiler
all the same. Updated init-asgs.sh to use master.