--to accepts comma or space delimited email address lists now
Issue 133: This change allows the --to option to utilize a list of
email addresses that are delimited by either spaces or commas. Since
this script is meant to be mostly a drop in replacement for sendmail,
it and sendmail delimits email addresses by spaces, it's important
to support this.

(cherry picked from commit 1182033bc8ec8ff25faf133fadfedf8dbba01692)
Signed-off-by: wwlwpd <6434714+wwlwpd@users.noreply.github.com>