Major update to asgs-brew.pl, major addition to interactive environment.
Issue #144: Created a shell wrapper that makes available the ASGS that
asgs-brew.pl makes available. Added some example "helper" to show what is
possible in the future.  Moved INSTALL to cloud/general, removed mention of
local::lib in the manual perlbrew installation instructions.

Issue #140: Removed installation of Date::Pcalc because it required
being treated differently than the other modules and its tests were
failing. Replaced all instances of Date::Pcalc in scripts with
Date::Calc. This commit does NOT remove the old version(s) of Date::Pcalc
that are distributed with ASGS, but that (and removing all CPAN
modules that are distributed) is the next step since the installation
of all non-custom modules is being handled using perlbrew and cpanm.

Issue #135: Adding better checks for failure, removed use of local::lib,
which seems to have been causing some confusion.

Issue #139: Added netcdf/hdf5 init scripts missing from a previous commit.

Issue #127: Also added a "setup-env" step and partitined out
the setting of specific environmental variables to the appropriate
step. Added a MACHINENAME variable that reflects the value of
the required --machinename flag..

Added step to install required Python libraries.

Issue #136: Covers pika, netCDF5, and the pptx modules. Did not verify
that RabbitMQ support is provided.

Issue #139: Split steps, NetCDF4 step is still dependant on HDF5; so
to enforce the precondition on the netcdf4 step is identical to the
postcondition check for hdf5.

Issue #138: Added options to OpenMPI's configure command. This reduced
build time significantly. Mostly it was the tracing libraries and tools
that were disabled.

Addition items not tied to a specific issue...

asgs-init.sh now asks a series of questions to create the asgs-brew.sh
command used because there was a need to source platforms.sh before,
therefore the need to properly bootstrap the environment was required
even for the initial running of asgs-brew.pl; this bootstrap script
can be distributed as a single point of installation for most systems
that are supported in platforms.sh; that is it can be distributed
alone and when run, checkouts the ASGS repo if not found - thus providing
access to all required files needed for set up

updated documentation

A refining, imported properties.sh for future access to run.properties
settings

Added sourcing steps for logging.sh and platforms.sh to asgsh generation
step.

Updated error messages to reflect supported subcommands; on start
the shell now imports the functions defined in platform.sh and
logging.sh; then it attempts to run 'env_dispatch' based on the
value of $MACHINENAME as set by asgs-brew.pl. This is still
experimental and no practical use has been devised. However, these
are essential to asge_main.sh, so any issues discovered by doing
this should be fixed in the shell rather than being avoided.

reload will reread the profile for a current session, thus re-parsing
the config file and looking for state file, log, and properties info.
save now works with no NAME param if session is currently loaded, calls
the reload routine if used with a currently loaded session; useful for
when one is first setting up a config or operations for a new storm

Using INSTANCENAME that can be grep'd out of the config file, we
can discover the location of the state file in the scratch dir; from
there we can discover the RUNDIR, SYSLOG, and location of the primary
run.properties file. Using this information, implemented 'watchlog'
that 'tail -f's the log file. No derived information is save in the
named profile since it should be discovered by first inspecting the
configuration file.

updated Vagrant file

added load/save/delete

make sure that all include Perl and Python scripts can pass
their respective compiliation phases; a byproduct is a set of
.pyc files, which are removed - but these could prove to be useful
at a later date. Update existing Perl scripts so they pass minimal
checks.