asgs-brew changes needed for the 2019stable release.
Issue #136: Added step to install python 2.7.17 to ensure
that the python environment is consistent among the various
platforms that must be supported. This was necessary because
a Perl module that ASGS requires was not able to get built
properly on some platforms due to the provide python providing
an incorrect openssl environment. Beyond that, it ensures more
stability.

Issue #161: Updated documentation and added an INSTALLING file
in the root ASGS directory describing how to quickly install
ASGS on a new platform.

Issue #162: Updated platforms.sh to be more consistent with
ADCIRC's cmplrflags.mk file and the specific platform environments
contained therein.