Making ASGS aware of $HOME/.ssh/config for ssh-based connections.
With the addition of an ssh config under the control of the operator,
the scripts need only to reference the Host alias as defined in
$HOME/.ssh/config. Under each stanza, any of the existing parameters
may be changed (Username, Port, etc) as well as any ssh or scp
options. Since rsync uses ssh by default, it would automatically
derive host specific settings by virtue of using the Host alias as
the reference to the remote server.

This change separates the concerns of ASGS from managing ssh usernames,
hosts, private keys, and "-o" options to focusing on using ssh connections
on a higher level for interacting with remote servers.

The change also allows operators to draw from the same basic ssh config
file, requiring them to only ensure that their respective usernames and
paths to private keys are correct. The list of servers and ssh options
can be maintained and be consistent across the entire effort and in
conjunction with resource administrators.

Resolves Issue #131.