Demonstration of how an ssh config would simplify ssh/scp/rsync
Issue #231: This is just a demonstration. With the addition of an
ssh config under the control of the operator, the scripts need only
to reference the Host alias as defined in $HOME/.ssh/config. Under
each stanza, any of the existing parameters may be changed
(Username, Port, etc) as well as any ssh or scp options. Since rsync
uses ssh by default, it would automatically derive host specific
settings by virtue of using the Host alias as the reference to
the remote server.