Fixed relocatability bug in asgs-brew.pl
Issue #171: some of the steps in asgs-brew.pl had references to
$HOME and broke when trying to install it into alternative locations.
This fix removes the fragility and should allow ASGS to be installed
anywhere, to anywhere. Access to a $USER's true $HOME is needed for
Perl, Python, and asgsh; since these 3 components are meant to serve
all of an operators installations globally.