Added ADCIRC build support to ASGS Shell and asgs-brew.pl
Issue #185: Updated ADCIRC build support in asgs-brew.pl by making
it optionally enabled using the --build-adcirc flag. Added
ADCIRC environment loading and management in the ASGS Shell
and added the 'initadcirc' command interface an interactive shell
script that guides the user to downloading the ADCIRC source
if needed through to the building of ADCIRC. The ADCIRC environment
then is made available for export into an asgsh session and
saving to an ASGS profile. Updated asgsh's 'list' and 'load'
commands to support augmenting an active asgsh's environment with
the exported ADCIRC related variables associated with the build.
The most relevant to asgs_main.sh being ADCIRCDIR.

Documentation updates included reflecting changes to asgs-brew.pl
in the internal POD (viewed using perldoc) and to the file describing
the ASGS Shell, cloud/general/README.asgsh.