Fixed file ID and changed behavior for exact coordinate matches
ASCII output was sent to file id 61 in one line, instead of it being dynamically based on the open file.  More substantially, if the station coordinates exactly match the coordinates of a node (determined based on the weight being 1d0), then wet/dry state of the element is ignored and the matching node's values are output directly.  I think this is consistent with what ADCIRC's internal station output does.