Adding initial Dockerfile
Issue #165: Adding a Dockerfile to support the evaluation,
exploration, and local use of the 2019stable release of ASGS.
The Dockerfile is used by a developer or maintainer when after
a release they wish to make a Docker image of the environment
available at https://hub.docker.com.

As part of this effort, 2019stable is available on a Ubuntu based
Docker image via:

  https://hub.docker.com/r/asgsdockerhub/2019stable

To use, someone would:

1. install Docker on their local laptop or workstation
2. run, docker pull asgsdockerhub/2019stable
3. run, docker run -it asgsdockerhub/2019stable

This will start the container in the ASGS Shell environment; to check
things are properly configured, run the 'verify' command at the asgsh
prompt.

NOTE 1:

ADCIRC source nor binaries are distributed in this image. The "initadcirc"
command is available to assist someone wishing to obtain and build ADCIRC,
but they will need access to the source somehow.

NOTE 2:

This change also includes a minor change to the script that installs
the Perl modules because during the development of the Dockerfile,
it was discovered that the upstream version of Date::Handler had some
failing tests that were preventing the module from being installed.
This was through no fault of ASGS, nor does it appear that this test
failing affects ASGS in any negative way.