Allow advisoryMonitor.sh to fallback to http (non-ssl) for testing.
Issue #209: Added code to fallback to http (non-ssl) when there is
a 599 HTTP response code, indicating in this case that the connection
to port 443 (https) was refused. Cleaned up a few things while in
there; the first 4 commandline arguments are preserved, but added
a few more to allow for specifying FTP hostname/ip, RSS hostname/ip,
trigger type, etc. Any other cleaning up preserved the existing
functionality.

This was mainly done to make it easier to test ad hoc FTP and WWW
servers that might be set up to mimic the NHC infrastructure, thus
providing another and more realistic way to do integration testing
(and real investigation/research).