Added tools for easily replaying storms.
Issue #215: This is the initial script for replaying storms
in a configurable way for testing, development, training,
conducting readiness drills. There are lots of opportunities
to determine how to extend this script. It is meant to run on
a server that is configured with FTP and RSS (via http) which
mimicks how the NHC provides advisories during actual events.
Documentation for the script is provided via POD and can be
viewed using "perldoc ./replay-storm.pl".

In this first version, index-at.xml files are updated for each
storm by adding an entry for each storm. Storms are updated in
place, but the existing scheme allows each storm among many to
exist the XML exactly one time.