Created "nowification" script. Added SSL to replay nginx.
Issue #216: The main addition in this commit is the advisory_nowify.pl,
which is meant move old storms to a new start time. This is mainly for
testing and drill purposes, and the generated advisory files are meant
to be served up by the replay script. This commit also includes some
updates to the replay script and some tools to make it easier to add
SSL to the nginx container that serves up the RSS advisory feeds.