Fixed get_atcf.pl so it can handle index-at.xmls with > 1 storm.
Issue #221: Introduced XML::RSS into get_atcf.pl, but didn't change
the way it parses the file except for how it gets the <description>
body out of the XML. Before the file is processed, XML::RSS is used
to modify the forecast file so that the wanted storm (based on $storm
and $year) is the only <item> in the XML. So it basically filters the
rest out. The XML as it would have been had $storm $year been the only
storm is passed to the existing processing. The modification persists
in the index-at.xml that is written to disk for consumption by other
utilities that may be expecting a single storm contained in the XML.