Retargeted, Updated to Python 2.7.18. Deleted noisy Python test script.
Issue #226: Python 2.7 environment now updates into path specified via
--install-path ($ASGS_INSTALL_PATH). Prior to this change, it was installing
into $HOME regardless of what the install path was. This was affecting users'
quotas in $HOME, which tend to be set low. Most want to install in their
$WORK directories also. Python 2.7.18 was released in April, so updated the
version installed. This has no real effect other than the comfort of tracking
the "latest" 2.7 (since LTS has ended). Removed a test that was throwing
a unicode error. It was a test script from UNIDATA testing the netCDF module,
but it's okay to remove since the issue was NOT netCDF related and there
is another script that exercises the Python netCDF module.