Avoid spurious module warnings on Hatteras when starting asgsh.
Issue #241: When starting asgsh, platforms.sh file has a case
statement in it for Hatteras that was issuing a warning if the
current user was not accounted for; removed this warning and moved
module includes inside of the specific users. Set default modules
to include for unspecified users to what seemed same - specifying
the Intel compiler suite and OpenMPI module.

For the existing users on Hatteras accounted for in this way,
care was taken to ensure the effective environment that was being
configured remained the same as before this change.

Auto-close link: Resolves Issue #241.