Replaced 'set' command with 'define' due to bash keyword conflict.
Issue #244: This issue was discovered when implementing commandline
flag processing to the asgsh shell command that starts the ASGS
Shell Environment. 'set' is a built-in keyword of bash and asgsh
was using it for other purposes.  The new command is 'define' and
works exactly the same as 'set' did. The documentation included
with asgsh was also updated to reflect this change.

Currently supported flags for asgsh:

-h             -> print command line synopsis, then quits to login shell
-p profileName -> set profile name to start in at the commandline  (e.g.,: % asgsh -p storm042020)

Added one more "fix" not tracked by an issue, but when defining the
config file, files specified using a relative path are converted to
their absolute path form before being saved to the environment.

Resolves Issue #244.