Pinned pip module versions and cleaned up python installation script.
Issue #252: To ensure a compatible environment for Python 2.7, versions
for the modules installed via pip were pinned. This was most important
for the netCDF4 module, which removed support for Python 2.7. netCDF4
1.5.2 is also pinned to numpy 1.16.6, so that was also enforced at
the pip level when installing it explicitly. Internal changes to
asgs-brew.pl make it also a little easier to maintain version upgrades.
Even though upstream says there are no more Python 2.7 updates, they
a new minor version was released in March. It's highly likely they
will continue with minor changes like this from time to time to address
security issues and critical bugs, since 2.7 is likely to be heavily
used for many years to come due to it's large installation base.

Resolves #252.