Removed reliance on RMQMessaging_Python envar, purged platforms.sh
Issue #258: This is a rather aggressive change, but one that is needed
for python stability. It removes the need for "RMQMessaging_Python" to
be set and is not used to run any scripts since all .py scripts should
have the proper shebang and be executable as a result of #256. The loading
of Python via the environment tool "module" that is common on HPCs has
also been removed to further minimize any confusion about which `python`
should be used to execute any scripts.

Resolves #258.