Added regession test for monitoring/FortCheck.py
Issue #260: Added a smallish fort.61.nc for FortCheck.py to read;
this is a regression test to ensure that the Python environment in
the ASGS Shell Environment (asgsh) is working as expected.

If the test fails, it indicates a problem with how the environment
finds the right python executable to run or indicates issues with
the installation of one of the Python modules.

Also cleaned up the output of the 'verify' command so it's easier
to differentiate the different pieces of the over all testing. We
hope to make adding regression testing (even a "long test") part
of the SOP.

Resolves #260.