Cleaned up the basis of the 'initadcirc' command in asgsh.
Issue #267: Improved consistency in guiding questions/options with
./init-asgs.sh (e.g., "'.'" to skip the 'git checkout' part when
as an alternative to specifying a branch - useful for development
environments when the git branch of ADCIRC (or ASGS) may be in an
unclean (non-committed) state.

This was not necessarily a bug, but was noticed during the preparation
of an ASGS Shell video overview as being inconsistent.

Resolves #267.