Cleaning up exported paths for openmpi (gfortran only) and ADCIRC
Issue #290: added conditionals to export_ENV hashes inside of the steps
for openmpi, which is only built currently if the --compiler=gfortran
is used; also selectively setting some ADCIRC_* environmental variables,
but exposing them all so that they may be saved. If all are blocked, then
what asgsh knows to save to a profile is incomplete.

WIP, need to test on supermic.