Cleaning up exported paths for openmpi (gfortran only) and ADCIRC
Issue #290: added conditionals to export_ENV hashes inside of the steps
for openmpi, which is only built currently if the --compiler=gfortran
is used; also selectively setting some ADCIRC_* environmental variables,
but exposing them all so that they may be saved. If all are blocked, then
what asgsh knows to save to a profile is incomplete.

Also added a verify check (verify_netcdf) that checks for HDF5 and NetCDF
binaries and libraries in expected locations. Binaries should be available
in PATH and are checked via `which $binary`; libraries should be located
in $ASGS_INSTALL_PATH/lib, and are checked directly to exist using -e.

Resolves #290 - or at least provides for some assurances that the environment
is working as expected.