Added 'edit statefile', 'inspect' alias for 'edit', 'guess platform'
Issue #286: Added 'statefile' to 'edit' command; also added an alias
for 'edit' named, 'inspect' so that it's semantically correct and may
be easier for the user to recall in a pinch.

Resolves #286

Issue #294: In this commit is an unrelated change that updates 'sq'
so that it the correct command is called for the current platform
for listing jobs that $USER currently had in the queue - in any state
(waiting, running, etc). This required the addition of a "guess platform"
command that may server as the basis for a number of enhancements,
which require knowing what machine we're running on.

Resolves #294