Added an interactive 'clone profile' operation in asgsh.
Issue #296: 'clone profile' allows the operator to quickly start
a new profile or set of runs using an existing profile that may
only need to change in configuratino by instance name. When run,
the operator is run through a series of questions that result in
a new configuration file, new profile, and chance to edit the new
configuration file.

Resolves #296.