Disabled SSL verification in get_atcf.pl for use with test server.
There is no issue attached to this, but the change is necessary to make
get_atcf.pl accept https servers that do not have valid SSL files, which
is needed for testing using the replay server. This makes the script
more permissive and will not have an impact on it during a real event
as it retrieves files from the NHC site.

Also in this update are some necessary changes to the formatting of the
XML files generated by the nowification script. Their liberal use of
newlines XML::RSS's "save" and "as_string" methods was breaking some
assumptions in get_atcf.pl used in the iterative parsing (line by
line) that is done there.

Also moved replay-storm.ini to a "sample" file so that the one actually
used is not tracked with changes in git.