Fully relocated perl and python 2.7 from $HOME to --install-path.
This change moves the rest of python 2.7 to the directory specified
by --install-path (aka $ASGS_INSTALL_PATH in asgsh). It also moves
the installation of perlbrew, perl, and required perl modules out
of $HOME and into --install-path.  There are many benefits to this,
the biggest 2 are that it doesn't affect the $HOME quota which is very
common on shared HPC resources and it makes the ASGS installation
var more isolated. For example, ASGS' perlbrew maintained perl env-
ironment and python's use of pip to install modules can easily
interfere with a user that chooses to use either of them with the
default "home" paths set. Additionally, scripts used by asgs-brew.pl
now use /tmp/asgs-$USER as the temp directory. It's protected with
`chmod 700`, but this was done to further remove file clutter from
$HOME that is not under the user's control.

Resolves #338.
Resolves #339.
Resolves #340.