Enforcing consistency for setting of WORK, SCRATCH, SCRATCHDIR
Issue #282: Added setting of WORK and SCRATCH in init-asgs.sh; fixed
non-capture of WORK and SCRATCH in asgsh profile if set; works with
init-asgs.sh's setting of WORK and SCRATCH in the environment. Builds
on expectation that ASGS is installed via init-asgs.sh + asgs-brew.pl.

Removed determination of SCRIPTDIR out of default platforms.sh for most
current platforms, resting on SCRATCH being used by asgs_main.sh to
derive this if not set.

Resolves #282.