Changing default ADCIRC_GIT_URL url to https, updated docs.
No issue created for this, but the change is necessary so that when
'initadcirc' is run, it prompts for username and password. If an
operator wishes to use the ssh form of the git url, they will need
to update the environmental variable ADCIRC_GIT_URL to use the url,

   export ADCIRC_GIT_URL=git@github.com:adcirc

in the profile they're in when they run 'initadcirc' inside of asgsh.
This is being done to make it easier to install on fresh installs where
setting up an ssh key on github is super inconvenient. If someone is
building ADCIRC enough to want to avoid entering in their GH username/
password can work around this using the method described above.