Added 'list jobs' command.
Issue #379: Extracts all queue job Ids associated with a profile
if RUNDIR is defined and exists. The intended purpose of this is to
support a forced shutdown of an ASGS instance so that all job Ids
can be cancelled be they running, waiting to run, or having already
been run (no-op).

Resolves #379.