Use different backtrace option based on compiler.
Issue 393: Intel compilers were not able to find the "backtrace"
routine, rather it needs the "tracebackqq" routine - gfortran does
use the "backtrace" routine. A define flag was added and is set
when the it is detected that "ifort" is being used to compile all
of Fortran code (or most of it) in ASGS.

The change to use "tracebackqq" is only activated when the Intel
compiler is used, for all others the original "backtrace" is used.
This has only been tested to work with Intel compilers.

This issue was blocking Issue 385 (ImageMagick) and further development
efforts.

Resolves Issue 393.