Addressed issues on Hatteras for Imagic and more "tracebackqq" fixes.
This commit addresses issues that were exposed on Hatteras regarding
ImageMagick (specified CC=gcc in our install script) and found a few
more places where we had to specify "-DINTEL" so that "tracebackqq"
was used rather than "backtrace" in the "logging.mod" Fortran module
library.

Added a step in asgs-brew.pl that preconditions the shell environment
in which all the building is happening. This was necessary because we
found that some Perl module builds will fail if any of the following
variables are present initially: PERL5LIB, PERL_MB_OPT, PERL_LOCAL_LIB_ROOT,
and PERL_MM_OPT - this moves further to the idea of providing a fully
isolated environment inside of the ASGH Shell.

Renamed the helper wrapper created on the initial build from "update-asgsh"
to "update-asgs", which is really what the tool does (i.e., provides
a wrapper around the asgs-brew.pl command containing the initial flags.)

Also specified arch as "native" for the ImageMagick "configure" script
that is called so that it'd work on Lonestar5 - seems to work fine on
the other machines test.

Resolves Issue 385.
Resolves Issue 393.