Updated Dockerfiles to include zip and remove gnuplot-nox.
Issue 389: This issue came down to making sure 'zip' was available
on Ubuntu based Docker images, so the 'fix' was to just reconcile
the Dockerfiles to this need. 'zip' is available on all RedHat based
machines - which all of the HPC platforms we use are. So it makes sense
to take care of the preconditions on Ubuntu in the Dockerfiles them-
selves.

Resolves Iusse 389.