Added support build and "profile" for ADCIRC v55release.
Issue #391 - Adding support for building v55 of ADCIRC and
updating support for the "load adcirc" command (which updates
$PATH, etc based on where the binaries are expected to be upon
successful compilation)

Change summary:

* treats an upstream (the private repo) branch of ADCIRC as the "version"
* using this info, looks for the ./swan directory in the right place (it
  was moved to ./thirdparty/swan since v55, before that it was a top level
  directory at the same level as 'work'
* updated how "load adcirc <adcirc-profile-name>" puts the binary locations
  in PATH; originally it created symlinks in $ASGS_INSTALL_PATH/bin, but
  this prevented the same ASGS installation from using different "versions"
  (branch builds) of ADCIRC; now the asgsh command simply prepends PATH
  with the appropriate directories (i.e., work/, thirdparty/swan)

Support of v55 or other ADCIRC "versions" should base the determination of
behavor on the value of $ADCIRC_GIT_BRANCH, which is effectively what
we wish to take rather than "version". So there is no "version" variable,
but there is $ADCIRC_GIT_BRANCH. For example, this is used in the script
that provides the functionality of the "initadcirc" asgsh command in this
commit in the following ways:

a) it requires that only a 'supported' branch may be specified
b) it uses the value of ADCIRC_GIT_BRANCH to decide where it expects to
   find the ./swan directory (as mentioned above).

NOTE: This commit doesn't represent an exhaustive validation process
for use of v55 in ASGS operationally. But this commit is a
precondition for such a validation.

Resolves #391.