Added workaround for ImageMagick build issue, first noticed on Debian.
Issue 417: PerlMagick doesn't correctly determine the location of libperl;
on systems where there is no system perl or libdev doesn't exist at all, the
build fails. On systems where it does, the wrong one is used (via system perl).
These patches and changes to the building of ImageMagick fix this on all tested
platforms.

A Dockerfile for Debian (stretch-slim) is also provided.

Resolve #417.