Added nco v4.9.8 to list of supported programs.
Issue 402: Installed support for The netCDF Operators (NCO) toolkit
within asgs-brew.pl. This includes: an entry in asgs-brew.pl, an
installation script that supports a "--clean" flag; and when building
fetches the specified version from Github, unarchives it, and runs
the ./configure/make/make install steps.

The initial version of nco supported is 4.9.8, "Death Valley".

Resolves #402.