Made default ADCIRC_GIT_URL ssh based.
Issue #455: Github is deprecating their https URIs for their repos,
particularly impacting private repos. ADCIRC is contained in a private
repository.

Https was being used originally because it presented to the ASGS user
the familiar git username/password prompt. So it is now unavoidable
that a user must set up their ssh keys in Github.

GH has decent documentation, so the impact will be softened but this
fact. The link to their document on this topic is,

    https://docs.github.com/en/github/authenticating-to-github/adding-a-new-ssh-key-to-your-github-account

Resolves #455.