Adding patch to support gfortran to build swan in v55release.
Issue #424: Added patch set to switch the default macros.inc for swan
to use gfortran, rather than intel compilers. This is needed on Ubuntu
and any other platform where gfortran is the only compiler available.

Adjusted references to v55release internally to the commit SHA
representation, which can be treated just like a branch or tag.

Resolves #424.