Don't "repatch" ADCIRC if already applied once. Select by number option added.
Issue #479: When patches are applied for the first time, a 'touch' file is created
in the ADCIRC source directory that indicates the branch/sha, date/time, and patch
path for each patch file. Subsequent builds of this version check for the 'touch' file,
and skip patching if it's detected.

Issue #480: Added and have generally tested "select by number" in the ADCIRC build
menu.

Resolves #479. Resolves #480.