Expanded ASGS env variables save by asgsh shell profiles.
Issue 504: Now saving the following as part of profile:

    *HINDCASTWALLTIME
    *ADCPREPWALLTIME
    *NOWCASTWALLTIME
    *FORECASTWALLTIME
    *QUEUENAME
    *SERQUEUE
    *ACCOUNT
    *PPN
    *INTENDEDAUDIENCE
    *USERIVERFILEONLY
    *RIVERSITE
    *RIVERDIR
    *RIVERUSER
    *RIVERDATAPROTOCOL
    *FTPSITE

In addition to just saving these in the asgsh profile, asgs_main.sh
is also aware that they may be set in the environment. As a result,
when the following variables are set in the environment (managed by
the profile), those values are used in lieu of the existing defaults.

Resolves #504.