Updated references to the main upstream repo.
Issue 514: Updates included some changes to Dockerfiles that were
still pointing to the old upstream. Not all references were update,
particular those in input files and configuration files. This may
be addressed at a later time, but for now is not critical.

No functional changes were made in this commit outside of the rarely
used Dockerfiles.

Resolves #514.