Added support for upstream ADCIRC tag, v55.00.
Issue 524: Note v55.00 is an "upstream" git tag and not the same
as we've come to know "v55release", which was a branch and is now
gone. The commit SHA remains.

Adding this version is to bring us back in line the line of active
"stable" ADCIRC support.

Resolves #524.

(cherry picked from commit 383071ef6e6ae0ea26fe56f71a3bf3d7aed3b4bb)
Signed-off-by: wwlwpd <46434714+wwlwpd@users.noreply.github.com>