Add initial TROPICALCYCLONE=monitor support.
Issue 271: Added STORM, STORMNAME (if in config file), and YEAR
to run.properties if TROPICALCYCLONE=monitor.

This approach doesn't use get_atcf.pl to proactively get
information about an active storm, but the plan is do this.

Also made it so if the rmq stuff if "off", this is reflected but
the rest of the variables related are not.

Resolves #271.