Made findAndClearOrphans Docker env friendly.
Issue 473: Updated 'comm' to be 'command' so the filtered output
was able to be checked for 'tail -f /dev/null', which is used
by the docker-entryscript.sh to keep the container "alive". So,
this commit restores the functionality removed in the previous
commit, but in a way that leave the parent docker container proc
alone.

Resolves #473.