Localized "$THIS" where possible, inside of bash function.
Issue 573: Added 'local' keyword to assignment to "$THIS" in all
bash functions. When used globally, set "THIS=(basename -- $0)",
which provides consistent naming of "$THIS" across all scripts.

Furthermore, there's no need to reset or "change back" $THIS when
a function is called. This took care of itself when a script that
defined $THIS globally in the script file exited because a child
process can't affect the environment of the parent.

Resolves #573.