Commit 2 of 3 - replace -O3 with -O2 in ADCIRC 53release.
Issue 549: Added patch set for v53release that changes -O3 to -02
on top of the base patch set already presented for v53release.

This change includes a modification to the 'build adcirc' script
in the way that the menu is build so that it's more efficient to
manage this menue in future changes.