Commit 3 of 3 - replace -O3 with -O2 in ADCIRC 55release.
Issue 549: Added patch set for v53release that changes -O3 to -02
on top of the base patch set already presented for v53release.

This change includes update to the 'build adcirc' script and the
fix to address the misplaced 'endif' in the queenbeeC patch (01)
that has been applied to the existing v55release* patch sets in
another commit.

Resolves #549.