Addressing issue with stdout redirected to 'null' in logMessage.
Issue-570: There was a trailing white space in 'findAndClearOrphans'
that likely I introduced not long ago, so deleted that. Also updated
the check in logMessage to ensure that not only did the file exist
(would be true for a directory also), but that it existed *and*
was a file (as opposed to it being a directory).

Resolves #570 (hopefully :-) ).