Audited transmit_rps.sh, things requires RMQMessaging_Enable=on
Issue 584: Audited transmit_rps.sh; ensured that RMQ* things
(including setting/exporting of environmental variables *and*
inclusion in the run.properties files) were minimized if
RMQMessaging_Enable=off (or rather, not "on").

In addition to this, made $RMQMessaging_Enable and $RMQMessaging_Transmit
may be set in the configuration file and be respected when the
asgs_main.sh script initializes default values.

Also, it was found that in platforms.sh, all major HPC platforms
were being forcibly setting RMQMessaging_Enable and $RMQMessaging_Transmit
to "on", despite the fact that they were "off" by default in the
asgs_main.sh script. This was changed to respect the defaults that
are set in asgs_main.sh (and can be affected in the configuration
file) for all HPC platforms execpt for hatteras.

This turned into follow up to the "rmq-audit" branch.

Resolves #584.