Made `$TMP` directory used to build various things relocatable.
Issue 624: On systems where /tmp is not a suitable base directory
for downloading upstream source code for compiling, this change
allows whomever is building ASGS to define a TMPDIR when running
init-asgs.sh. In addition to this, all platforms except for hatteras
default to /tmp/$USER-asgs; on hatteras it defaults to use a
directory in $WORK.

Furthermore, if TMPDIR is set in the environment, that is used as
the base temp directory rather than /tmp.

Resolves #624.