Made the get ATCF script relocatable, defaults to get_atcf.pl
Issue #635: Added the ability to define the get ATCF script via
configuration file using $GET_ATCF_SCRIPT. This is so that the
script that monitors and retrieves the NHC forecasts and best
track data can be easily swapped by setting this variable in a
configuratin file. This is mainly a tool for developing new
get ATCF script, but there might be other uses as well.

Resolves #635.